Add bot greeting response

+++ case_files/src/models/chat_model.rs
@@ -2,7 +2,8 @@ use serde::{ Deserialize, Serialize };
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Chat {
-    id: i32,
+    id: i64,
+    #[serde(rename = "type")]
     _type: ChatType,
     title: Option<String>,
     username: Option<String>,
@@ -18,17 +19,18 @@ pub struct Chat {
     join_by_request: Option<bool>,
     description: Option<String>,
     invite_link: Option<String>,
-    slow_mode_delay: Option<i32>,
-    message_auto_delete_time: Option<i32>,
+    slow_mode_delay: Option<i64>,
+    message_auto_delete_time: Option<i64>,
     has_aggressive_anti_spam_enabled: Option<bool>,
     has_hidden_members: Option<String>,
     has_protected_content: Option<bool>,
     sticker_set_name: Option<String>,
     can_set_sticker_set: Option<bool>,
-    linked_chat_id: Option<i32>,
+    linked_chat_id: Option<i64>,
 }
 
 #[derive(Debug, Deserialize, Serialize)]
+#[serde(rename_all = "lowercase")]
 pub enum ChatType {
     Private,
     Group,
