+++ case_files/src/models/contact_model.rs
@@ -5,6 +5,6 @@ pub struct Contact {
     phone_number: String,
     first_name: String,
     last_name: Option<String>,
-    user_id: Option<i32>,
+    user_id: Option<i64>,
     vcard: Option<String>,
 }
